chore: service, controller name rollback
