+++ backend/src/main/java/com/NBE4_5_SukChanHoSu/BE/domain/movie/controller/MovieContentsController.java
@@ -1,0 +1,48 @@
+package com.NBE4_5_SukChanHoSu.BE.domain.movie.controller;
+
+import com.NBE4_5_SukChanHoSu.BE.domain.movie.dto.request.MovieRequest;
+import com.NBE4_5_SukChanHoSu.BE.domain.movie.entity.Movie;
+import com.NBE4_5_SukChanHoSu.BE.domain.movie.service.MovieContentsService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/movie")
+public class MovieContentsController {
+
+    private final MovieContentsService movieContentsService;
+
+    @PostMapping
+    public ResponseEntity<Movie> createMovie(@RequestBody MovieRequest request) {
+        Movie movie = request.toEntity(); // MovieRequest → Movie 변환
+        Movie saved = movieContentsService.save(movie);
+        return ResponseEntity.ok(saved);
+    }
+
+    @GetMapping
+    public ResponseEntity<List<Movie>> getAllMovies() {
+        return ResponseEntity.ok(movieContentsService.findAll());
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<Movie> getMovieById(@PathVariable Long id) {
+        return movieContentsService.findById(id)
+                .map(ResponseEntity::ok)
+                .orElse(ResponseEntity.notFound().build());
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Movie> updateMovie(@PathVariable Long id, @RequestBody Movie movie) {
+        return ResponseEntity.ok(movieContentsService.update(id, movie));
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteMovie(@PathVariable Long id) {
+        movieContentsService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+}
