+++ backend/src/main/java/com/NBE4_5_SukChanHoSu/BE/domain/movie/service/MovieContentsService.java
@@ -1,0 +1,47 @@
+package com.NBE4_5_SukChanHoSu.BE.domain.movie.service;
+
+import com.NBE4_5_SukChanHoSu.BE.domain.movie.entity.Movie;
+import com.NBE4_5_SukChanHoSu.BE.domain.movie.repository.MovieRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+public class MovieContentsService {
+
+    private final MovieRepository movieRepository;
+
+    public Movie save(Movie movie) {
+        return movieRepository.save(movie);
+    }
+
+    public List<Movie> findAll() {
+        return movieRepository.findAll();
+    }
+
+    public Optional<Movie> findById(Long movieId) {
+        return movieRepository.findById(movieId);
+    }
+
+    public Movie update(Long movieId, Movie updatedMovie) {
+        Movie movie = movieRepository.findById(movieId)
+                .orElseThrow(() -> new IllegalArgumentException("Movie not found: " + movieId));
+
+        movie.setGenres(updatedMovie.getGenres());
+        movie.setTitle(updatedMovie.getTitle());
+        movie.setReleaseDate(updatedMovie.getReleaseDate());
+        movie.setPosterImage(updatedMovie.getPosterImage());
+        movie.setDescription(updatedMovie.getDescription());
+        movie.setRating(updatedMovie.getRating());
+        movie.setDirector(updatedMovie.getDirector());
+
+        return movieRepository.save(movie);
+    }
+
+    public void delete(Long movieId) {
+        movieRepository.deleteById(movieId);
+    }
+}
